Fix spelling in raw scopes
Previously the contents of a raw scope included an unused capture group (2) that caused the contents to be tokenized into multiple tokens, which prevented the spell checker from being run on the whole last word.

For instance ````Test```` would be tokenized to `Tes` and `t`. This is because the final `t` was captured by group 2. Since captures are normally used to apply different scope names, it had to be split into its own token.

By converting it to a non-capturing group (since it isn't used in `captures`), the whole contents of the raw section can be tokenized properly and then spell-checked properly.